feat: separate stacinfo and tools widget
