+++ widgets/EodashStacInfo.vue
@@ -1,33 +1,4 @@
 <template>
-  <div class="d-flex flex-column fill-height overflow-hidden">
-    <div
-      v-if="showHeader"
-      class="d-flex flex-row flex-shrink-0 justify-space-between pa-4 bg-primary align-center"
-    >
-      <v-btn
-        v-if="showIndicatorsBtn"
-        color="secondary"
-        class="text-none py-2 px-4"
-        :append-icon="[mdiPlus]"
-        text="Choose indicator"
-        @click="dialog = !dialog"
-      >
-        Replace indicator
-      </v-btn>
-      <EodashLayoutSwitcher v-if="showLayoutSwitcher" target="main" />
-    </div>
-    <PopUp v-model="dialog" maxWidth="1000px" width="1000px">
-      <EodashItemFilter
-        .enableCompare="true"
-        imageProperty="assets.thumbnail.href"
-        subTitleProperty="subtitle"
-        resultType="cards"
-        .filterProperties="[]"
-        filtersTitle=""
-        resultsTitle="Explore more indicators"
-        @select="dialog = !dialog"
-      />
-    </PopUp>
     <div class="flex-grow-1 fill-height overflow-auto">
       <eox-stacinfo
         .for="currentUrl"
@@ -42,23 +13,13 @@
       >
       </eox-stacinfo>
     </div>
-  </div>
 </template>
 
 <script setup>
 import "@eox/stacinfo";
-import PopUp from "./PopUp.vue";
-import EodashItemFilter from "./EodashItemFilter.vue";
-import EodashLayoutSwitcher from "./EodashLayoutSwitcher.vue";
 import { currentUrl } from "../core/client/store/States";
-import { mdiPlus } from "@mdi/js";
-import { ref } from "vue";
-
-const dialog = ref(false);
 
 const {
-  showIndicatorsBtn,
-  showLayoutSwitcher,
   allowHtml,
   featured,
   footer,
@@ -68,14 +29,6 @@ const {
   subheader,
   tags,
 } = defineProps({
-  showIndicatorsBtn: {
-    type: Boolean,
-    default: true,
-  },
-  showLayoutSwitcher: {
-    type: Boolean,
-    default: true,
-  },
   allowHtml: {
     type: Boolean,
     default: true,
@@ -125,5 +78,4 @@ main {padding-bottom: 10px;}
     default: () => ["sci:citation"],
   },
 });
-const showHeader = showIndicatorsBtn || showLayoutSwitcher;
 </script>
